Moved formatTime() and formatDate() to utils.js
